requirements.txt
numpy
onnxruntime
pandas==2.2.3
matplotlib==3.8.1
seaborn==0.13.2
tqdm


gymnasium<1.0.0
moviepy
protobuf
pygame
stable-baselines3
wandb
torchrl
tensordict
tyro